Fix missing client address on Node v0.10
At the time when the response was sent, the client address is no longer
available.

This commit fixes the problem by saving the address at the time when
the request has arrived.

Since we are already modifying the `req` object, we can save the start
time there too and avoid a closure variable.